+++ Makefile
@@ -1,25 +1,25 @@
-CC		  = gcc
-CFLAGS	  = -g -Wall -Wextra -std=c11 -MMD -Wmissing-prototypes -Wmissing-declarations -Wshadow -Wmissing-variable-declarations -fno-common
+CC          = gcc
+CFLAGS      = -g -Wall -Wextra -std=c11 -MMD -Wmissing-prototypes -Wmissing-declarations -Wshadow -Wmissing-variable-declarations -fno-common
 UNITY_DIR   = ../Unity
 TEST_CFLAGS = -I$(UNITY_DIR)/src -I.
-TEST_SRC	= test/test_venn3.c test/test_d6.c test/test_initialize.c test/test_known_solution.c test/test_main.c \
-              test/test_graphml.c test/test_venn4.c test/test_venn5.c  test/test_venn6.c 
-TEST_BIN	= $(TEST_SRC:test/%.c=bin/%)
+TEST_SRC    = test/test_venn3.c test/test_s6.c test/test_initialize.c test/test_known_solution.c test/test_main.c \
+              test/test_graphml.c test/test_venn4.c test/test_venn5.c test/test_venn6.c
+TEST_BIN    = $(TEST_SRC:test/%.c=bin/%)
 # Do not include entrypoint.c in the test builds, it contains the main function, which is also in the test files.
-SRC		    = main.c failure.c color.c cycle.c cycleset.c edge.c log.c vertex.c statistics.c s6.c face.c dynamicface.c utils.c \
-			  memory.c graphml.c triangles.c engine.c corners.c initialize.c nondeterminism.c innerface.c venn.c save.c
-TEST_HELPERS = test/test_helpers.c
-XSRC		= entrypoint.c
-HDR			= color.h cycle.h cycleset.h dynamicface.h edge.h statistics.h core.h face.h main.h trail.h \
-			  s6.h failure.h vertex.h memory.h common.h triangles.h engine.h nondeterminism.h
-OBJ3		= $(SRC:%.c=objs3/%.o) $(TEST_HELPERS:test/test_%.c=objs3/test_%.o)
-OBJ4		= $(SRC:%.c=objs4/%.o) $(TEST_HELPERS:test/test_%.c=objs4/test_%.o)
-OBJ5		= $(SRC:%.c=objs5/%.o) $(TEST_HELPERS:test/test_%.c=objs5/test_%.o)
-OBJ6		= $(SRC:%.c=objs6/%.o)
-TEST_OBJ6 = $(TEST_HELPERS:test/test_%.c=objs6/test_%.o)
-XOBJ		= $(XSRC:%.c=objs6/%.o)
-DEP		 = $(OBJ6:.o=.d) $(OBJ5:.o=.d) $(OBJ4:.o=.d) $(OBJ3:.o=.d) $(XOBJ:.o=.d) $(TEST_SRC:test/%.c=bin/%.d)
-TARGET	  = bin/venn
+SRC         = main.c failure.c color.c cycle.c cycleset.c edge.c log.c vertex.c statistics.c s6.c face.c dynamicface.c utils.c \
+              memory.c graphml.c triangles.c engine.c corners.c initialize.c nondeterminism.c innerface.c venn.c save.c
+TEST_HELPERS = test/helper_for_tests.c
+XSRC        = entrypoint.c
+HDR         = color.h cycle.h cycleset.h dynamicface.h edge.h statistics.h core.h face.h main.h trail.h \
+              s6.h failure.h vertex.h memory.h common.h triangles.h engine.h nondeterminism.h
+OBJ3        = $(SRC:%.c=objs3/%.o) $(TEST_HELPERS:test/%.c=objs3/%.o)
+OBJ4        = $(SRC:%.c=objs4/%.o) $(TEST_HELPERS:test/%.c=objs4/%.o)
+OBJ5        = $(SRC:%.c=objs5/%.o) $(TEST_HELPERS:test/%.c=objs5/%.o)
+OBJ6        = $(SRC:%.c=objs6/%.o)
+TEST_OBJ6   = $(TEST_HELPERS:test/%.c=objs6/%.o)
+XOBJ        = $(XSRC:%.c=objs6/%.o)
+DEP         = $(OBJ6:.o=.d) $(OBJ5:.o=.d) $(OBJ4:.o=.d) $(OBJ3:.o=.d) $(XOBJ:.o=.d) $(TEST_SRC:test/%.c=bin/%.d)
+TARGET      = bin/venn
 
 .SECONDARY: 
 
@@ -57,7 +57,7 @@ bin/test_%: objst/test_%.o $(UNITY_DIR)/src/unity.c $(OBJ6) $(TEST_OBJ6)
 .format: $(SRC) $(HDR) $(TEST_SRC) $(XSRC) $(D6) $(TEST_HELPERS)
 	clang-format -i $?
 	for f in $?; do \
-		if [ $$(tail -c 1 "$$f" | od -An -t x1) != "0a" ]; then \
+		if [ $$(tail -c 1 "$$f" | od -An -t x1) \!= "0a" ]; then \
 			echo >> "$$f"; \
 		fi; \
 	done
@@ -77,26 +77,26 @@ objsv/test_venn%.o: test/test_venn%.c
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=$(*F) -c $< -o $@
 
-objst/test_%.o: test/test_%.c
+objst/%.o: test/%.c
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=6 -c $< -o $@
 
-objs3/test_%.o: test/test_%.c
+objs3/%.o: test/%.c
 	@echo Compiling $<
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=3 -c $< -o $@
 
-objs4/test_%.o: test/test_%.c
+objs4/%.o: test/%.c
 	@echo Compiling $<
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=4 -c $< -o $@
 
-objs5/test_%.o: test/test_%.c
+objs5/%.o: test/%.c
 	@echo Compiling $<
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=5 -c $< -o $@
 
-objs6/test_%.o: test/test_%.c
+objs6/%.o: test/%.c
 	@echo Compiling $<
 	@mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(TEST_CFLAGS) -DNCOLORS=6 -c $< -o $@
